Moved global const, type, and var into globals.go
